requirements.txt
requests>=2.32.0
pydantic>=2.7.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
typer>=0.12.3
rich>=13.7.1
